[FEAT] mypage 앱 생성
