+++ chatting/models.py
@@ -1,7 +1,7 @@
 from django.db import models
-from users.models import *
 from mentoring.models import *
 from django.utils import timezone
+from users.models import *
 
 class Chatroom(models.Model):
     mentor = models.ForeignKey(Mentor, on_delete=models.CASCADE)
@@ -11,17 +11,10 @@ class Chatroom(models.Model):
 
 class ChatInterest(models.Model):
     chatroom = models.ForeignKey(Chatroom, on_delete=models.CASCADE)
-    interest = models.ForeignKey(Interest, on_delete=models.CASCADE)
+    interest = models.ForeignKey('users.Interest', on_delete=models.CASCADE)
 
 class Chat(models.Model):
     room = models.ForeignKey(Chatroom, on_delete=models.CASCADE)
     message = models.CharField(max_length=150)
     is_mentee = models.BooleanField(default=False)
-    created_at = models.DateTimeField(default=timezone.now)
-
-class Log(models.Model):
-    author = models.ForeignKey(User, on_delete=models.CASCADE)
-    chatroom = models.ForeignKey(Chatroom, on_delete=models.CASCADE)
-    title = models.CharField(max_length=30)
-    content = models.TextField()
     created_at = models.DateTimeField(default=timezone.now)
