+++ mentoring/serializers.py
@@ -4,35 +4,6 @@ from django.contrib.auth import get_user_model
 from .models import *
 from users.serializers import *
 
-# User = get_user_model()
-
-# class UserSerializer(serializers.ModelSerializer):
-#     class Meta:
-#         model = User
-#         fields = ['id', 'username', 'name']
-
-# class InterestSerializer(serializers.ModelSerializer):
-#     class Meta:
-#         model = Interest
-#         fields = ['name']
-
-# class MentorSerializer(serializers.ModelSerializer):
-#     user = UserSerializer()
-#     interests = InterestSerializer(many=True)
-
-#     class Meta:
-#         model = Mentor
-#         fields = ['id', 'user', 'interests', 'rating']
-
-# class MenteeSerializer(serializers.ModelSerializer):
-#     user = UserSerializer()
-#     # interests = InterestSerializer(many=True)
-#     # 멘티가 관심설정한 멘토 목록
-
-#     class Meta:
-#         model = Mentee
-#         fields = ['user', 'interests']
-
 class ConcernSerializer(serializers.ModelSerializer):
     author = MenteeSerializer(read_only=True)
     interests_ids = InterestSerializer(many=True, read_only=True)
