+++ mentoring/views.py
@@ -3,8 +3,8 @@ from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework import viewsets, status
 from django.contrib.auth import login, logout, authenticate
-# from users.serializers import *
-# from users.models import *
+from users.serializers import *
+from users.models import *
 from .serializers import *
 from .models import *
