+++ mypage/admin.py
@@ -1,5 +1,4 @@
 from django.contrib import admin
 from .models import *
 
-admin.site.register(Chatroom)
-admin.site.register(Chat)
+admin.site.register(Log)
